Se corrige metodos que generaban a errror

+++ controller/control.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from gc import set_debug
 import tkinter
 import model.client as client_model
 import model.database as db
@@ -9,9 +8,10 @@ import view.gui as view
 class Controller :
 
     def __init__(self):
+        """Se inicializa la clase con la creacion de objetos en los modulos modelo y vista"""
         self.control_db = db.Database("Hotel Teressita")
         self.control_hotel_model = hotel_model.Hotel_model()
-        self.create_table()
+        self.create_table() 
         self.root = tkinter.Tk()
         self.control_view = view.Interface(self.root,self)
         
@@ -31,27 +31,28 @@ class Controller :
             );'''
         self.control_db.create_table(query)
 
-    def create_client(self):
-        entry = self.control_hotel_model.date_model(self.control_view.entry_date.get())
-        exit = self.control_hotel_model.date_model(self.control_view.exit_date.get())
-        self.client = client_model.Client(self.control_view.name_client.get(),self.control_view.last_name.get(),
-                self.control_view.dni.get(),self.control_view.room.get(),entry,exit)
-        query = """--sql
-            INSERT INTO Clientes
-            values(NULL, ?, ?, ?, ?, ?, ?);
-            """
-        parameters = (self.client.get_last_name(), self.client.get_last_name(), 
-                self.client.get_dni(),self.client.get_room(), self.client.get_entry_date(), 
-                self.client.get_exit_date())
+    def create_client(self,client_data):
         try : 
+            entry = self.control_hotel_model.date_model(client_data['entry_date'])
+            exit = self.control_hotel_model.date_model(client_data['exit_date'])
+            self.client = client_model.Client(client_data['name'],client_data['last_name'],
+                    client_data['dni'],client_data['room'],entry,exit)
+            query = """--sql
+                INSERT INTO Clientes
+                values(NULL, ?, ?, ?, ?, ?, ?);
+                """
+            parameters = (self.client.get_name(), self.client.get_last_name(), 
+                    self.client.get_dni(),self.client.get_room(), self.client.get_entry_date(), 
+                    self.client.get_exit_date())
             self.control_db.create(query, parameters)
+            self.control_view.set_variables()
             return ("Crear Cliente",
                     "EL cliente fue creado correctamente")
         except :
             return ("Crear Cliente",
                     "Hubo un error, el cliente no fue guardado")
     
-    def read_clients(self,tree):
+    def read_clients(self,tree,ob=''):
         clients = tree.get_children()
         for client in clients:
             tree.delete(client)
@@ -63,13 +64,9 @@ class Controller :
             tree.insert("", "end", text=key,
                      values=(data[key][0], data[key][1], data[key][2],
                              data[key][3], data[key][4], data[key][5]))
-    
-    def query_client(self,tree):
-        
-        clients = tree.get_children()
-        for client in clients:
-            tree.delete(client)
 
+    def query_client(self,tree,client_data,entry_date,exit_date):
+        search_client = False
         search_name = search_last_name = search_dni = search_room = "%%"
         search_entry_date = "0000-01-01"
         search_exit_date = "9999-12-31"
@@ -83,48 +80,56 @@ class Controller :
             AND fechaDESalida <= ?
             ORDER BY apellido ASC;
             """
-        if len(self.control_view.name_client.get()) != 0:
-                search_name = "%" + self.control_view.name.get() + "%"
-        if len(self.control_view.last_name.get()) != 0:
-                search_last_name = "%" + self.control_view.last_name.get() + "%"
-        if len(self.control_view.dni.get()) != 0:
-                search_dni = "%" + self.control_view.dni.get() + "%"
-        if self.control_view.room.get() != "Seleccionar":
-                search_room = self.control_view.room.get()
-        if len(self.control_view.entry_date.get()) != 0:
-                search_entry_date = self.control_hotel_model.date_model(self.control_view.entry_date)
-        if len(self.control_view.entry_date.get()) != 0:
-                search_exit_date = self.control_hotel_model.date_model(self.control_view.exit_date)
-        parametros = (search_name, search_last_name, search_dni, search_room, 
-                search_entry_date,search_exit_date)
-        data = self.control_db.read(query, parametros) 
-        if len(data) == 0:
-            return ("Consultar","El cliente no existe")
+        if len(client_data['name']) != 0:
+            search_client = True
+            search_name = "%" + client_data['name'] + "%"
+        if len(client_data['last_name']) != 0:
+            search_client = True
+            search_last_name = "%" + client_data['last_name'] + "%"
+        if len(client_data['dni']) != 0:
+            search_client = True
+            search_dni = "%" + client_data['dni'] + "%"
+        if client_data['room'] != "Seleccionar":
+            search_client = True
+            search_room = client_data['room']
+        if len(client_data['entry_date']) != 0:
+            search_client = True
+            search_entry_date = self.control_hotel_model.date_model(client_data['entry_date'])
+        if len(client_data['exit_date']) != 0:
+            search_client = True
+            search_exit_date = self.control_hotel_model.date_model(client_data['entry_date'])
+        parameters = (search_name, search_last_name, search_dni, search_room, 
+                    search_entry_date,search_exit_date)
+        data = self.control_hotel_model.clients_data_estructure(self.control_db.read(query,parameters))
+        if len(data) == 0 or not search_client:
+            self.control_view.message_box(("Consultar","El cliente no existe") if search_client else ("Consultar","Ingrese un valor a buscar"))
         else:
-            self.control_hotel_model.clients_data_estructure(data)
+            clients = tree.get_children()
+            for client in clients:
+                tree.delete(client)
             for key in data.keys():
                 tree.insert("", "end", text=key,
                      values=(data[key][0], data[key][1], data[key][2],
                              data[key][3], data[key][4], data[key][5]))
     
-    def update_client(self,id,name,last_name,dni,room,entry_date,exit_date):
-        entry = self.control_hotel_model.date_model(entry_date)
-        exit = self.control_hotel_model.date_model(exit_date)
-        self.client = client_model.Client(name.get(),last_name.get(),
-                dni.get(),room.get(),entry_date,exit_date)
-        query = """--sql
-            UPDATE Clientes
-            SET nombre = ?,
-                apellido = ?, 
-                DNI = ?,
-                habitacion = ?,
-                fechaDeIngreso = ?,
-                fechaDeSalida = ?
-            WHERE ID = ?;"""
-        parameters = (self.client.get_last_name, self.client.get_last_name, 
-                self.client.get_dni,self.client.get_room, self.client.get_entry_date, 
-                self.client.get_exit_date,id)
+    def update_client(self,client_data):
         try:
+            entry = self.control_hotel_model.date_model(client_data['entry_date'])
+            exit = self.control_hotel_model.date_model(client_data['exit_date'])
+            self.client = client_model.Client(client_data['name'],client_data['last_name'],
+                    client_data['dni'],client_data['room'],entry,exit)
+            query = """--sql
+                UPDATE Clientes
+                SET nombre = ?,
+                    apellido = ?, 
+                    DNI = ?,
+                    habitacion = ?,
+                    fechaDeIngreso = ?,
+                    fechaDeSalida = ?
+                WHERE ID = ?;"""
+            parameters = (self.client.get_name(), self.client.get_last_name(), 
+                    self.client.get_dni(),self.client.get_room(), self.client.get_entry_date(), 
+                    self.client.get_exit_date(),client_data['id_client'])
             self.control_db.create(query, parameters)
             return ("Modificar Cliente",
                     "El cliente fue modificado correctamente")
@@ -132,19 +137,27 @@ class Controller :
             return ("Crear Cliente",
                     "Hubo un error, el cliente no fue modificado")
 
-    def delete_client(self,id):
-        query = """--sql
-            DELETE FROM Clientes
-            WHERE ID = ?;"""
-        parameters = (id,)
-        try:
-            self.control_db.delete(query, parameters)
-            return ("Borrar Cliente",
-                    "EL cliente fue borrado correctamente")
-        except :
-            return ("Borrar Cliente",
-                    "Hubo un error,EL cliente no fue borrado")
+    def delete_client(self):
+        client = self.control_view.tree.item(self.control_view.tree.focus())
+        self.control_view.id_client = client['text']
+        if self.control_view.id_client !='':
+            if self.control_view.message_box():
+                query = """--sql
+                    DELETE FROM Clientes
+                    WHERE ID = ?;"""
+                parameters = (self.control_view.id_client,)
+                try:
+                    self.control_db.delete(query, parameters)
+                    self.control_view.message_box(("Borrar Cliente","EL cliente fue borrado correctamente"))
+                    self.read_clients(self.control_view.tree)
+                except :
+                    self.control_view.message_box(("Borrar Cliente",
+                            "Hubo un error,EL cliente no fue borrado"))
+        else :
+            self.control_view.message_box(("Borrar Cliente","Seleccione el cliente que quiere eliminar"))
+        self.control_view.set_variables()
     
+    """El siguiente metodo cumple con el control de la habitaciones del hotel este igualmente tiene un modelo que es fue estableciodo en el modulo modelo"""
     def avalible_rooms(self,event,variable_button,room_form,entry_date,exit_date):
         if variable_button.get() == "Buscar":
             room_form["values"] = self.control_hotel_model.get_hotel_rooms()
@@ -174,42 +187,47 @@ class Controller :
             room_form["values"] = self.control_hotel_model.avalible_rooms(data)
             return room_form
 
+    """Los siguientes metodos dan control a validez de las strings y enteros que se ingresen en la aplicación se utiliza un modelo con el modulo regex"""
     def validate_string(self, text):
         return self.control_hotel_model.validate_string(text)
     
     def validate_number(self, text):
         return self.control_hotel_model.validate_number(text)
 
-    def action_press(self, action):
-        if action == 'Guardar':
-            print(action)
-            self.create_client()
-
-        elif action == 'Clientes':
-            #self.read_clients()
-            print(action)
-            self.control_view.variable_button.set("guudfudar")
-        elif action == 'Buscar':
-            #self.query_client()
-            print(action)
-            self.control_view.variable_button.set("guudfudar")
-        elif action == 'Actualizar':
-            #self.update_client()
-            print(action)
-            self.control_view.variable_button.set("guudfudar")
-        elif action ==  'Borrar':
-            #self.delete_client()
-            print(action)
-
-    def validate_date(self):
+    def validate_data(self,name_error,last_name_error,dni_error,room_error,client_data):
+        self.control_view.set_labels()
         validate = True
-        if not self.control_hotel_model.validate_string(self.control_view.name_client.get()):
-            self.control_view.name_error.set("Ingrese el nombre")
+        if not self.control_hotel_model.validate_string(client_data['name'],'^[a-zA-ZÀ-ÿ\u00f1\u00d1\s]+$'):
+            name_error.set("Ingrese el nombre")
             validate = False
-        if not self.control_hotel_model.validate_string(self.control_view.last_name.get()):
-            self.control_view.last_name_error.set("Ingrese el apellido")
+        if not self.control_hotel_model.validate_string(client_data['last_name'],'^[a-zA-ZÀ-ÿ\u00f1\u00d1\s]+$'):
+            last_name_error.set("Ingrese el apellido")
             validate = False
-        if not self.control_hotel_model.validate_number(self.control_view.dni.get(),"^[0-9]{8}$"):
-            self.control_view.dni_error.set("Ingrese un DNI valido")
-        if self.control_view.room.get() == "Seleccionar":
-            self.control_view.room_error.set("Sin Seleccionar")
+        if not self.control_hotel_model.validate_number(client_data['dni'],'^[0-9]{8}$'):
+            dni_error.set("Ingrese un DNI valido")
+            validate = False
+        if client_data['room'] == "Seleccionar":
+            room_error.set("Sin Seleccionar")
+            validate = False
+        return validate
+    
+    """el siguiente metodo cumple con el control del boton que se situa en el frame de los formularios, 
+    se controla el crear un clienmte, buscar un cliente o actualizar un cliente """    
+    def action_press(self,client_data):
+        if self.control_view.variable_button.get() == 'Guardar':
+            if self.validate_data(self.control_view.name_error,self.control_view.last_name_error,\
+                self.control_view.dni_error,self.control_view.room_error,client_data):
+                result = self.create_client(client_data)
+                self.read_clients(self.control_view.tree)
+                self.control_view.message_box(result)
+        elif self.control_view.variable_button.get() == 'Buscar':
+            self.query_client(self.control_view.tree,client_data,self.control_view.entry_date,self.control_view.exit_date)
+        elif self.control_view.variable_button.get() == 'Actualizar':
+            if self.validate_data(self.control_view.name_error,self.control_view.last_name_error,\
+                self.control_view.dni_error,self.control_view.room_error,client_data):
+                result = self.update_client(client_data)
+                self.read_clients(self.control_view.tree)
+                self.control_view.message_box(result)
+   
+    
+
